@@ -11,20 +11,21 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^11.0.5",
-    "@angular/common": "^11.0.5",
-    "@angular/compiler": "^11.0.5",
-    "@angular/core": "^11.0.5",
-    "@angular/forms": "^11.0.5",
-    "@angular/platform-browser": "^11.0.5",
-    "@angular/platform-browser-dynamic": "^11.0.5",
-    "@angular/router": "^11.0.5",
-    "bootstrap": "3.3.7",
+    "@angular/animations": "13.0.2",
+    "@angular/common": "12.2.13",
+    "@angular/compiler": "12.2.13",
+    "@angular/core": "12.2.13",
+    "@angular/forms": "12.2.13",
+    "@angular/platform-browser": "12.2.13",
+    "@angular/platform-browser-dynamic": "12.2.13",
+    "@angular/router": "12.2.13",
+    "@popperjs/core": "^2.10.2",
+    "bootstrap": "^5.1.3",
     "bulma": "^0.9.3",
-    "rxjs": "^6.6.3",
-    "rxjs-compat": "^6.0.0",
-    "tslib": "^2.0.3",
-    "zone.js": "^0.10.3"
+    "rxjs": "7.4.0",
+    "src": "^1.1.2",
+    "tslib": "2.3.1",
+    "zone.js": "0.11.4"
   },
   "devDependencies": {
     "@angular-devkit/build-angular": "^0.1100.4",
