@@ -1,13 +1,20 @@
-import { AfterViewInit, Component, ElementRef, OnDestroy, OnInit, ViewChild } from '@angular/core';
+import {
+  AfterViewInit,
+  Component,
+  ElementRef,
+  OnDestroy,
+  OnInit,
+  ViewChild,
+} from '@angular/core';
 import { fromEvent, Subscription } from 'rxjs';
-import { debounceTime, map, mergeAll, tap } from 'rxjs/operators'
-import { PostService } from 'src/app/services/post.service';
+import { debounceTime, map, mergeAll, tap } from 'rxjs/operators';
+import { PostService } from '../../services/post.service';
 import { post } from '../../obj/post';
 
 @Component({
   selector: 'app-game',
   templateUrl: './game.component.html',
-  styleUrls: ['./game.component.scss']
+  styleUrls: ['./game.component.scss'],
 })
 export class GameComponent implements OnInit, AfterViewInit, OnDestroy {
   posts: post[] = [];
@@ -18,41 +25,44 @@ export class GameComponent implements OnInit, AfterViewInit, OnDestroy {
 
   @ViewChild('quoteInput') input: ElementRef<any> = new ElementRef(null);
 
-  constructor(public postService: PostService) { }
+  constructor(public postService: PostService) {}
 
   ngOnInit(): void {
-    this.subscribtions.push(this.postService.get('https://jsonplaceholder.typicode.com/posts')
-      .pipe(
-        mergeAll()
-      )
-      .subscribe(data => this.posts.push(data)))
+    this.subscribtions.push(
+      this.postService
+        .get('https://jsonplaceholder.typicode.com/posts')
+        .pipe(mergeAll())
+        .subscribe((data) => this.posts.push(data))
+    );
   }
 
   ngAfterViewInit(): void {
     fromEvent<any>(this.input.nativeElement, 'keyup')
       .pipe(
-        map(data => data.target.value),
-        tap(data => console.log(data)),
+        map((data) => data.target.value),
+        tap((data) => console.log(data)),
         debounceTime(50),
-        map(data => this.phrase = data)
+        map((data) => (this.phrase = data))
       )
-      .subscribe(_ => this.checkPhrase())
+      .subscribe((_) => this.checkPhrase());
   }
 
   onStart() {
-    this.quote = this.posts[this.postService.randomNum()].title
+    this.quote = this.posts[this.postService.randomNum()].title;
   }
 
   onFinish() {
-    if (this.phrase === this.quote){
-      alert('Complimenti hai vinto, premere start per una nuova partita')
+    if (this.phrase === this.quote) {
+      alert('Complimenti hai vinto, premere start per una nuova partita');
     } else {
-      alert('Peccato sembra che tu abbia fatto degli errori, premere start per riprovare')
+      alert(
+        'Peccato sembra che tu abbia fatto degli errori, premere start per riprovare'
+      );
     }
-    this.input.nativeElement.value = ''
+    this.input.nativeElement.value = '';
   }
 
-  checkPhrase(){
+  checkPhrase() {
     this.counter = 0;
     var arr1 = this.phrase.split(' ');
     var arr2 = this.quote.split(' ');
@@ -62,7 +72,7 @@ export class GameComponent implements OnInit, AfterViewInit, OnDestroy {
         console.log(this.counter);
       }
     }
-    if(this.counter == arr2.length) {
+    if (this.counter == arr2.length) {
       this.onFinish();
     }
   }
